Size of type test.

+++ src/control/frequency.h
@@ -11,7 +11,7 @@ ulong frequency_get_nominal_freq();
 ulong *frequency_get_freq_rank_list();
 ulong frequency_set_all_cpus(ulong freq);
 ulong frequency_pstate_to_freq(uint pstate);
-ulong frequency_freq_to_pstate(uint freq);
+uint frequency_freq_to_pstate(ulong freq);
 void frequency_set_performance_governor_all_cpus();
 void frequency_set_userspace_governor_all_cpus();
